vdW-DF: Removing __SPIN_BALANCED pre-compiler option
When vdW-DF was first developed in 2004 (10.1103/PhysRevLett.92.246401),
a spin-polarized version of the functional was not available. As such,
an approximative spin-balancing scheme was implemented that allowed for
spin-polarized calculations with this non-spin-polarized functional.
With the development of the fully spin-polarized svdW-DF in 2015
(10.1103/PhysRevLett.115.136402), this approximation was no longer
necessary but was kept for testing. We are now removing the option to
compile with this approximative scheme. The corresponding pre-compiler
flag was used mostly in Modules/xc_vdW_DF.f90, but there was one
occurrence in PP/src/xc_vdW_scale_mod.f90.